Delete unrequired entity and related logic ContractDtoFindContractualFolders
