+++ digital-repository-back/src/main/java/co/unicauca/digital/repository/back/domain/mapper/contract/IContractDtoFindContractualFoldersMapper.java
@@ -1,19 +1,19 @@
-package co.unicauca.digital.repository.back.domain.mapper.contract;
+// package co.unicauca.digital.repository.back.domain.mapper.contract;
 
-import co.unicauca.digital.repository.back.domain.dto.contract.response.ContractDtoFindContractualFoldersResponse;
-import co.unicauca.digital.repository.back.domain.model.contract.ContractDtoFindContractualFolders;
-import org.mapstruct.Mapper;
-import org.mapstruct.Mappings;
+// import co.unicauca.digital.repository.back.domain.dto.contract.response.ContractDtoFindContractualFoldersResponse;
+// import co.unicauca.digital.repository.back.domain.model.contract.ContractDtoFindContractualFolders;
+// import org.mapstruct.Mapper;
+// import org.mapstruct.Mappings;
 
-/**
- * Class that defines the mapper in charge of the O/R transition to a DTO
- * object.
- */
-@Mapper(componentModel = "spring")
-public interface IContractDtoFindContractualFoldersMapper {
-    @Mappings({
+// /**
+//  * Class that defines the mapper in charge of the O/R transition to a DTO
+//  * object.
+//  */
+// @Mapper(componentModel = "spring")
+// public interface IContractDtoFindContractualFoldersMapper {
+//     @Mappings({
 
-    })
-    ContractDtoFindContractualFoldersResponse toDtoFind(final ContractDtoFindContractualFolders contract);
+//     })
+//     ContractDtoFindContractualFoldersResponse toDtoFind(final ContractDtoFindContractualFolders contract);
 
-}
+// }
