+++ digital-repository-back/src/main/java/co/unicauca/digital/repository/back/domain/service/contract/IContractService.java
@@ -1,14 +1,11 @@
 package co.unicauca.digital.repository.back.domain.service.contract;
 
-import co.unicauca.digital.repository.back.domain.model.collection.Collection;
 import co.unicauca.digital.repository.back.domain.dto.contract.request.ContractDtoCreateRequest;
 import co.unicauca.digital.repository.back.domain.dto.contract.request.ContractDtoUpdateRequest;
 import co.unicauca.digital.repository.back.domain.dto.contract.response.ContractDtoCreateResponse;
 import co.unicauca.digital.repository.back.domain.dto.contract.response.ContractDtoFindResponse;
 import co.unicauca.digital.repository.back.global.response.PageableResponse;
 import co.unicauca.digital.repository.back.global.response.Response;
-
-import java.util.List;
 
 /**
  * Interface that allows defining the business operations to be carried out on
@@ -65,5 +62,26 @@ public interface IContractService {
      *         information about the outcome of the transaction.
      */
     Response<Boolean> deleteContract(final int id);
+
+    /**
+     * Service to get all contracts sort by SigningDate time descending
+     *
+     * @param pageNo   Pagination Page number
+     * @param pageSize Pagination Page size
+     * @return {@link Response} Response object for the service, which contains
+     *         information about the outcome of the transaction.
+     */
+    Response<PageableResponse<Object>> getContractualFoldersSortBySigningDate(int pageNo, int pageSize);
+
+    /**
+     * Service to get all contracts sort by SigningDate time descending
+     *
+     * @param filter Filter for the contractual folder
+     * @param search Word to search like
+     * @return {@link Response} Response object for the service, which contains
+     *         information about the outcome of the transaction.
+     */
+    Response<PageableResponse<Object>> getContractualFoldersByFilter(int pageNo, int pageSize, String filter,
+            String search);
 
 }
