+++ digital-repository-back/src/main/java/co/unicauca/digital/repository/back/domain/service/contract/impl/ListContractualFolders.java
@@ -1,103 +1,86 @@
-package co.unicauca.digital.repository.back.domain.service.contract.impl;
+// package co.unicauca.digital.repository.back.domain.service.contract.impl;
 
-import co.unicauca.digital.repository.back.domain.dto.contract.response.ContractDtoFindContractualFoldersResponse;
-import co.unicauca.digital.repository.back.domain.mapper.contract.IContractMapper;
-import co.unicauca.digital.repository.back.domain.mapper.contract.IContractDtoFindContractualFoldersMapper;
-import co.unicauca.digital.repository.back.domain.model.contract.Contract;
-import co.unicauca.digital.repository.back.domain.model.contract.ContractDtoFindContractualFolders;
-import co.unicauca.digital.repository.back.domain.repository.contract.IContractFilterAndSearchPatternRepository;
-import co.unicauca.digital.repository.back.domain.repository.contract.IContractRepository;
-import co.unicauca.digital.repository.back.domain.service.contract.IListContractualFolders;
-import co.unicauca.digital.repository.back.global.response.PageableResponse;
-import co.unicauca.digital.repository.back.global.response.Response;
-import co.unicauca.digital.repository.back.global.response.handler.ResponseHandler;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Sort;
-import org.springframework.stereotype.Service;
+// import co.unicauca.digital.repository.back.domain.dto.contract.response.ContractDtoFindContractualFoldersResponse;
+// import co.unicauca.digital.repository.back.domain.mapper.contract.IContractMapper;
+// import co.unicauca.digital.repository.back.domain.model.contract.Contract;
+// import co.unicauca.digital.repository.back.domain.repository.contract.IContractRepository;
+// import co.unicauca.digital.repository.back.domain.service.contract.IListContractualFolders;
+// import co.unicauca.digital.repository.back.global.response.PageableResponse;
+// import co.unicauca.digital.repository.back.global.response.Response;
+// import co.unicauca.digital.repository.back.global.response.handler.ResponseHandler;
+// import org.springframework.data.domain.Page;
+// import org.springframework.data.domain.PageRequest;
+// import org.springframework.data.domain.Sort;
+// import org.springframework.stereotype.Service;
 
-import java.util.List;
-import java.util.stream.Collectors;
+// import java.util.List;
+// import java.util.stream.Collectors;
 
-@Service
-public class ListContractualFolders implements IListContractualFolders {
+// @Service
+// public class ListContractualFolders implements IListContractualFolders {
 
-    /** Object to perform CRUD operations on the Contract entity */
-    private final IContractRepository contractRepository;
+//     /** Object to perform CRUD operations on the Contract entity */
+//     private final IContractRepository contractRepository;
 
-    /**
-     * Object to perform CRUD operations on the Contract filter and search pattern
-     */
-    private final IContractFilterAndSearchPatternRepository findByFilterAndSearchPattern;
+//     /** Mapping object for mapping the Contract */
+//     private final IContractMapper contractMapper;
 
-    /** Mapping object for mapping the Contract */
-    private final IContractMapper contractMapper;
+//     /** Constructor */
+//     public ListContractualFolders(IContractRepository contractRepository, IContractMapper contractMapper) {
+//         this.contractRepository = contractRepository;
+//         this.contractMapper = contractMapper;
+//     }
 
-    /** Mapping object for mapping the Contract when trying to filter by patterns */
-    private final IContractDtoFindContractualFoldersMapper iContractDtoFindContractualFoldersMapper;
+//     /**
+//      * @see IListContractualFolders#getContractualFoldersSortBySigningDate(int,int)
+//      */
+//     @Override
+//     public Response<PageableResponse<Object>> getContractualFoldersSortBySigningDate(int pageNo, int pageSize) {
+//         PageRequest pageRequest = PageRequest.of(pageNo, pageSize, Sort.by("signingDate").descending());
+//         Page<Contract> responsePage = contractRepository.findAll(pageRequest);
 
-    /** Constructor */
-    public ListContractualFolders(IContractRepository contractRepository,
-            IContractFilterAndSearchPatternRepository findByFilterAndSearchPattern, IContractMapper contractMapper,
-            IContractDtoFindContractualFoldersMapper iContractDtoFindContractualFoldersMapper) {
-        this.contractRepository = contractRepository;
-        this.findByFilterAndSearchPattern = findByFilterAndSearchPattern;
-        this.contractMapper = contractMapper;
-        this.iContractDtoFindContractualFoldersMapper = iContractDtoFindContractualFoldersMapper;
-    }
+//         List<Object> contractDtoFindContractualFolderResponse = responsePage.get()
+//                 .map(contract -> {
+//                     ContractDtoFindContractualFoldersResponse response = this.contractMapper
+//                             .toDtoFindContractualFolders(contract);
+//                     response.setSigningYear(contract.getSigningDate().getYear());
+//                     response.setModality(contract.getModalityContractType().getModality().getName());
+//                     response.setContractType(contract.getModalityContractType().getContractType().getName());
+//                     response.setVendor(contract.getVendor().getIdentification());
+//                     return response;
+//                 }).collect(Collectors.toList());
 
-    /**
-     * @see IListContractualFolders#getContractualFoldersSortBySigningDate(int,int)
-     */
-    @Override
-    public Response<PageableResponse<Object>> getContractualFoldersSortBySigningDate(int pageNo, int pageSize) {
-        PageRequest pageRequest = PageRequest.of(pageNo, pageSize, Sort.by("signingDate").descending());
-        Page<Contract> responsePage = contractRepository.findAll(pageRequest);
+//         PageableResponse<Object> response = PageableResponse.builder()
+//                 .data(contractDtoFindContractualFolderResponse)
+//                 .pageNo(responsePage.getNumber())
+//                 .pageSize(responsePage.getSize())
+//                 .totalElements(responsePage.getTotalElements())
+//                 .totalPages(responsePage.getTotalPages())
+//                 .last(responsePage.isLast())
+//                 .build();
 
-        List<Object> contractDtoFindContractualFolderResponse = responsePage.get()
-                .map(contract -> {
-                    ContractDtoFindContractualFoldersResponse response = this.contractMapper
-                            .toDtoFindContractualFolders(contract);
-                    response.setSigningYear(contract.getSigningDate().getYear());
-                    response.setModality(contract.getModalityContractType().getModality().getName());
-                    response.setContractType(contract.getModalityContractType().getContractType().getName());
-                    response.setVendor(contract.getVendor().getIdentification());
-                    return response;
-                }).collect(Collectors.toList());
+//         return new ResponseHandler<>(200, "Exitoso", "Exitoso", response).getResponse();
+//     }
 
-        PageableResponse<Object> response = PageableResponse.builder()
-                .data(contractDtoFindContractualFolderResponse)
-                .pageNo(responsePage.getNumber())
-                .pageSize(responsePage.getSize())
-                .totalElements(responsePage.getTotalElements())
-                .totalPages(responsePage.getTotalPages())
-                .last(responsePage.isLast())
-                .build();
+//     @Override
+//     public Response<PageableResponse<Object>> getContractualFoldersByFilter(int pageNo, int pageSize, String filter,
+//             String search) {
+//         PageRequest pageRequest = PageRequest.of(pageNo, pageSize);
+//         Page<ContractDtoFindContractualFoldersResponse> responsePage = contractRepository
+//                 .findByFilterAndSearchPattern(filter, search, pageRequest);
 
-        return new ResponseHandler<>(200, "Exitoso", "Exitoso", response).getResponse();
-    }
+//         List<Object> contractDtoFindContractualFolderResponse = responsePage.get()
+//                 .collect(Collectors.toList());
 
-    @Override
-    public Response<PageableResponse<Object>> getContractualFoldersByFilter(int pageNo, int pageSize, String filter,
-            String search) {
-        PageRequest pageRequest = PageRequest.of(pageNo, pageSize);
-        Page<ContractDtoFindContractualFolders> responsePage = findByFilterAndSearchPattern
-                .findByFilterAndSearchPattern(filter, search, pageRequest);
+//         PageableResponse<Object> response = PageableResponse.builder()
+//                 .data(contractDtoFindContractualFolderResponse)
+//                 .pageNo(responsePage.getNumber())
+//                 .pageSize(responsePage.getSize())
+//                 .totalElements(responsePage.getTotalElements())
+//                 .totalPages(responsePage.getTotalPages())
+//                 .last(responsePage.isLast())
+//                 .build();
 
-        List<Object> contractDtoFindContractualFolderResponse = responsePage.get()
-                .map(
-                        this.iContractDtoFindContractualFoldersMapper::toDtoFind)
-                .collect(Collectors.toList());
-
-        PageableResponse<Object> response = PageableResponse.builder()
-                .data(contractDtoFindContractualFolderResponse)
-                .pageNo(responsePage.getNumber())
-                .pageSize(responsePage.getSize())
-                .totalElements(responsePage.getTotalElements())
-                .totalPages(responsePage.getTotalPages())
-                .last(responsePage.isLast())
-                .build();
-
-        return new ResponseHandler<>(200, "Exitoso", "Exitoso", response).getResponse();
-    }
-}
+//         return new ResponseHandler<>(200, "Exitoso", "Exitoso", response).getResponse();
+//     }
+// }
